requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
colorama==0.4.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.1.1
python-telegram-bot==20.6
pytz==2025.2
qrcode==8.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
